Parse update history prices endpoint to float before saving

+++ Krakenbot/models/firebase_users.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import TypedDict
 from Krakenbot import settings
 from google.cloud.firestore_v1.base_query import FieldFilter
@@ -7,7 +8,7 @@ from Krakenbot.exceptions import NoUserSelectedException
 
 class PortfolioValues(TypedDict):
 	uid: str
-	value: float
+	value: Decimal
 	time: datetime
 
 class FirebaseUsers:
@@ -18,7 +19,7 @@ class FirebaseUsers:
 		else:
 			self.__user_doc = None
 
-	def update_portfolio_value(self, value: float, time: datetime):
+	def update_portfolio_value(self, value: Decimal, time: datetime):
 		'''
 		raise `NoUserSelectedException` if not initialised with uid
 		'''
@@ -28,7 +29,7 @@ class FirebaseUsers:
 		time = datetime(time.year, time.month, time.day, time.hour, tzinfo=time.tzinfo)
 		current_timestamp = time.timestamp()
 		portfolio = self.__user_doc.collection('portfolio').document(str(current_timestamp))
-		portfolio.set({ 'time': time, 'value': value })
+		portfolio.set({ 'time': time, 'value': float(value) })
 
 	def batch_update_portfolio(self, values: list[PortfolioValues]):
 		batch = settings.db_batch
@@ -38,7 +39,7 @@ class FirebaseUsers:
 			time = datetime(time.year, time.month, time.day, time.hour, tzinfo=time.tzinfo)
 			current_timestamp = time.timestamp()
 			user_portfolio = self.__users.document(value['uid']).collection('portfolio').document(str(current_timestamp))
-			batch.set(user_portfolio, { 'time': time, 'value': value['value'] })
+			batch.set(user_portfolio, { 'time': time, 'value': float(value['value']) })
 
 		batch.commit()
